9077: Move stream utilities into named folder
