+++ shared/src/business/useCases/elasticsearch/processStreamUtilities.js
