+++ shared/src/business/useCases/elasticsearch/processStreamUtilities.test.js
